third_party/blink/web_tests/platform/linux/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/cos.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] cos float16 0D scalar
  assert_array_approx_equals: test cos float16 property 0, expected -0.73876953125 +/- 0.001953125, expected -0.73876953125 but got 0
[FAIL] cos float16 1D constant tensor
  assert_array_approx_equals: test cos float16 property 0, expected -0.73876953125 +/- 0.001953125, expected -0.73876953125 but got 0.0078125
[FAIL] cos float16 1D tensor
  assert_array_approx_equals: test cos float16 property 0, expected -0.73876953125 +/- 0.001953125, expected -0.73876953125 but got 0
[FAIL] cos float16 2D tensor
  assert_array_approx_equals: test cos float16 property 0, expected -0.73876953125 +/- 0.001953125, expected -0.73876953125 but got 0
[FAIL] cos float16 3D tensor
  assert_array_approx_equals: test cos float16 property 0, expected -0.73876953125 +/- 0.001953125, expected -0.73876953125 but got 0
[FAIL] cos float16 4D tensor
  assert_array_approx_equals: test cos float16 property 0, expected -0.73876953125 +/- 0.001953125, expected -0.73876953125 but got 0
Harness: the test ran to completion.
